requirements.txt
scrapy==2.6.1
scancode-toolkit==30.1.0
psycopg2==2.9.3
mysql-connector-python==8.0.29